refactor(Inventory): Inventory
클린 아키텍처 적용을 위해 Domain Entity와 Jpa Entity 분리
