--- src/main/java/com/space/server/domain/inventory/service/implementation/InventoryDeleter.java
@@ -1,17 +1,0 @@
-package com.space.server.domain.inventory.service.implementation;
-
-import com.space.server.domain.inventory.domain.Inventory;
-import com.space.server.domain.inventory.domain.repository.InventoryRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class InventoryDeleter {
-
-  private final InventoryRepository inventoryRepository;
-
-  public void delete(Inventory inventory) {
-    inventoryRepository.delete(inventory);
-  }
-}
