--- src/main/java/com/space/server/domain/inventory/service/implementation/InventoryUpdater.java
@@ -1,16 +1,0 @@
-package com.space.server.domain.inventory.service.implementation;
-
-import com.space.server.domain.inventory.domain.Inventory;
-import org.springframework.stereotype.Service;
-
-@Service
-public class InventoryUpdater {
-
-  public void equip(Inventory inventory) {
-    inventory.equip();
-  }
-
-  public void unEquip(Inventory inventory) {
-    inventory.unEquip();
-  }
-}
